turn ResolutionFragment into a DialogFragment for reusing in NewDisplayMode action
