+++ app/src/main/res/values/strings.xml
@@ -14,6 +14,7 @@
     <string name="dpi">DPI</string>
     <string name="scale">Scale</string>
     <string name="apply">Apply</string>
+    <string name="cancel">Cancel</string>
     <string name="reset">Reset</string>
     <string name="success">Success</string>
     <string name="confirmation_dialog">Confirmation Dialog</string>
